backend/requirements.txt
flask
flask_sqlalchemy
flask_jwt_extended
flask_cors
psycopg2-binary
werkzeug==2.3.7
bcrypt
